+++ plugins/jsonparser/changelog.md
@@ -1,4 +1,4 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- Upgrade dependency: core to 1.2.0
+- Upgrade dependency: core to 1.2.1 and framework to 1.1.1