+++ plugins/otel/changelog.md
@@ -1,5 +1,4 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- First version cut 🚀
-- Feature: Support OTLP collector over HTTP or gRPC protocol.
+- Upgrade dependency: core to 1.2.1 and framework to 1.1.1